ollama-debate/requirements.txt
# Python dependencies for debate evaluation system

# LLM API clients
groq>=0.4.0  # Groq API client (primary)
requests>=2.31.0  # For Ollama, Together AI, Hugging Face APIs

# Data handling
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.65.0

# Optional: Jupyter notebook support (for analysis)
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
